+++ include/asm-generic/bitops/fls64.h
@@ -17,17 +17,17 @@
 #if BITS_PER_LONG == 32
 static __always_inline int fls64(__u64 x)
 {
-	__u32 h = x >> 32;
+	__u32 h = (__u32) (x >> 32);
 	if (h)
-		return fls(h) + 32;
-	return fls(x);
+		return fls((int)h) + 32;
+	return fls((int)x);
 }
 #elif BITS_PER_LONG == 64
 static __always_inline int fls64(__u64 x)
 {
 	if (x == 0)
 		return 0;
-	return __fls(x) + 1;
+	return __fls((int)x) + 1;
 }
 #else
 #error BITS_PER_LONG not 32 or 64
